feat: refactoring itemController from controllers methods
